add policy_rule_id to session event for easier debugging of policy manager rules (Comment #12644)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/hades/src@42998 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ src/com/untangle/node/web_cache/WebCacheEvent.java
@@ -24,7 +24,7 @@ public class WebCacheEvent extends LogEvent implements Serializable
         hitCount = missCount = bypassCount = systemCount = hitBytes = missBytes = policyId = 0;
     }
 
-    public WebCacheEvent(Long policyId,long hitCount,long missCount,long bypassCount,long systemCount,long hitBytes,long missBytes)
+    public WebCacheEvent(Integer policyId,long hitCount,long missCount,long bypassCount,long systemCount,long hitBytes,long missBytes)
     {
         this.hitCount = hitCount;
         this.missCount = missCount;
@@ -54,7 +54,7 @@ public class WebCacheEvent extends LogEvent implements Serializable
     public void setMissBytes( long missBytes ) { this.missBytes = missBytes; }
 
     public Long getPolicyId() { return(policyId); }
-    public void setPolicyId( Long policyId ) { this.policyId = policyId; }
+    public void setPolicyId( Integer policyId ) { this.policyId = policyId; }
 
     @Override
     public void compileStatements( java.sql.Connection conn, java.util.Map<String,java.sql.PreparedStatement> statementCache ) throws Exception
